Move ParallelQueue into router
